requirements.txt
appnope==0.1.4
asttokens==2.4.1
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
executing==2.1.0
filelock==3.17.0
fonttools==4.54.1
fsspec==2024.12.0
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytube==15.0.0
pytubefix==8.12.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.8.0
six==1.16.0
soundfile==0.12.1
soxr==0.5.0.post1
split-folders==0.5.1
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.68
ultralytics-thop==2.0.14
urllib3==2.2.3
wcwidth==0.2.13
wget==3.2
